requirements.txt
streamlit>=1.35.0
pandas>=2.2.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
tensorflow-cpu>=2.11.0
joblib>=1.3.0
